// lambda/friends/shared/types.ts
export interface APIGatewayEvent {
  httpMethod: string;
  path: string;
  pathParameters: { [key: string]: string } | null;
  queryStringParameters: { [key: string]: string } | null;
  headers: { [key: string]: string };
  body: string | null;
  requestContext: {
    requestId: string;
    authorizer?: {
      claims: {
        sub: string;
        email: string;
        username: string;
      };
    };
  };
  multiValueHeaders?: { [key: string]: string[] };
  multiValueQueryStringParameters?: { [key: string]: string[] };
  stageVariables?: { [key: string]: string } | null;
  isBase64Encoded: boolean;
  resource: string;
}